feat: 支持对url进行仓库维度限频#1539

+++ src/backend/common/common-ratelimiter/src/main/kotlin/com/tencent/bkrepo/common/ratelimiter/algorithm/TokenBucketRateLimiter.kt
@@ -35,10 +35,10 @@ import java.util.concurrent.TimeUnit
  * 单机令牌桶算法实现
  */
 class TokenBucketRateLimiter(
-    private val permitsPerSecond: Long,
+    private val permitsPerSecond: Double,
 ) : RateLimiter {
 
-    private val guavaRateLimiter = com.google.common.util.concurrent.RateLimiter.create(permitsPerSecond.toDouble())
+    private val guavaRateLimiter = com.google.common.util.concurrent.RateLimiter.create(permitsPerSecond)
 
     override fun tryAcquire(permits: Long): Boolean {
         try {
